fix: Add timestamp to db + refactor
