--- app/models/__init__.py
@@ -1,3 +1,0 @@
-from .user_model import UserModel
-
-crud_user_model = UserModel()
